pyproject.toml
[project]
name = "app1"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sepi", email = "me@sepi.me" }
]
dependencies = [
    "Flask>=3.0.2",
    "Frozen-Flask>=1.0.2",
    "markdown>=3.5.2",
    "python-frontmatter>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "app1:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
indent-width = 2

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app1"]
